DB workflow integration
Exporting the db using phpmyadmin and trying JSON as the text serialised format.

After I thought about it for a second I realized we really do need text serialised format for versioning, because if people are working on wordpress at the same time the database will be accessed constantly and with a binary file it would be impossible to merger them without losing one of the branches’ work.