Trying mysqldump for the ultimate in database versioning
hope this is the one!

I ticked the allow-keywords, and order-by-primary thinking it might help keep things in order even if there are edits to the db, therefor enhancing the fidelity of this method for versioning. Seems okay in theory but will try it without these options to test especially if the changes I do in a second effect more things in the code than they should.